requirements.txt
aiogram
python-dotenv
motor
python-dateutil
